Switch back to simpler strategy when matching multi-segment needles - let each needle segment match haystack entirely
This is less strict and will let more false positives in but because
of how heatmaps are computed the original strategy is not intuitive
any more. E.g. extension (esp. first character after ‘.’) has
not-trivial weight so matchings tends to pick it and thus everything
from first matched character up to extension was conidered as match
and filtered out leaving segments nothing to work with.